renaming some files to match test.ts pattern
