+++ test/extension-tests/multi-root-UI/test/configure-and-build.test.ts
@@ -7,7 +7,7 @@ import {
   expect,
   getFirstSystemKit,
   getMatchingSystemKit
-} from '@test/util';
+} from '@test/helpers/util';
 import * as fs_ from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
