+++ test/extension-tests/multi-root-UI/test/debugger.test.ts
@@ -1,4 +1,4 @@
-import {DefaultEnvironment, expect, getFirstSystemKit} from '@test/util';
+import {DefaultEnvironment, expect, getFirstSystemKit} from '@test/helpers/util';
 import * as vscode from 'vscode';
 import CMakeTools from '@cmt/cmake-tools';
