+++ test/extension-tests/multi-root-UI/test/variant-envs.test.ts
@@ -1,7 +1,7 @@
 
 import * as api from '@cmt/api';
 import {CMakeCache} from '@cmt/cache';
-import {clearExistingKitConfigurationFile, DefaultEnvironment, expect, getFirstSystemKit} from '@test/util';
+import {clearExistingKitConfigurationFile, DefaultEnvironment, expect, getFirstSystemKit} from '@test/helpers/util';
 import {fs} from '@cmt/pr';
 import * as path from 'path';
 import * as vscode from 'vscode';
