+++ test/extension-tests/single-root-UI/test/envs-presets.test.ts
@@ -1,6 +1,6 @@
 import * as api from '@cmt/api';
 import {CMakeCache} from '@cmt/cache';
-import {DefaultEnvironment, expect} from '@test/util';
+import {DefaultEnvironment, expect} from '@test/helpers/util';
 import * as vscode from 'vscode';
 
 suite('[Environment Variables in Presets]', async () => {
